# docs/swagger.yaml
basePath: /api
definitions:
...
  GAMERS-BE_internal_discord_application_dto.DiscordChannel:
    properties:
      guild_id:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      position:
        type: integer
      type:
        type: integer
    type: object
  GAMERS-BE_internal_discord_application_dto.DiscordGuild:
    properties:
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        type: boolean
      permissions:
        type: string
    type: object
...
  GAMERS-BE_internal_global_response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gamers.com
    name: API Support
  description: API Server for GAMERS platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GAMERS API
  version: "1.0"
paths:
...
  /discord/guilds:
    get:
      consumes:
      - application/json
      description: Returns all guilds where both the GAMERS bot and the authenticated
        user are members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/GAMERS-BE_internal_discord_application_dto.DiscordGuild'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Get available guilds for contest creation
      tags:
      - Discord
  /discord/guilds/{guildId}/channels:
    get:
      consumes:
      - application/json
      description: Returns all text channels in a guild where both the bot and the
        authenticated user are members
      parameters:
      - description: Discord Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/GAMERS-BE_internal_discord_application_dto.DiscordChannel'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Get guild's text channels
      tags:
      - Discord
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"